Renamed MatchFoundController
